+++ OpenDentBusiness/TableTypes/CovCat.cs
@@ -10,9 +10,9 @@ namespace OpenDentBusiness{
 		///<summary>Description of this category.</summary>
 		public string Description;
 		///<summary>Default percent for this category. -1 to skip this category and not apply a percentage.</summary>
-		public long DefaultPercent;
+		public int DefaultPercent;
 		///<summary>The order in which the categories are displayed.  Includes hidden categories. 0-based.</summary>
-		public long CovOrder;
+		public int CovOrder;
 		///<summary>If true, this category will be hidden.</summary>
 		public bool IsHidden;
 		///<summary>Enum:EbenefitCategory  The X12 benefit categories.  Each CovCat can link to one X12 category.  Default is 0 (unlinked).</summary>
